Fix time display bug (#207)
* Change pub_date datatype to be bigint

* Change Message pubdate type to be long

* Add DTO models for frontend

* Use new models at frontend

* Change naming and place for new viewmodels

* Add extension method to reduce conversion code duplication

* Remove "deprecated" todos

* Remove DisplayFormat attribute

+++ MinitwitSimulatorAPI/Models/MessageDTO.cs
@@ -1,3 +1,3 @@
 namespace MinitwitSimulatorAPI.Models;
 
-public record MessageDTO(string Content, int PubDate, string User);
+public record MessageDTO(string Content, long PubDate, string User);
